BIG UPDATE
1. Got rid of some libaries and made my player class for playing music
P. S. code does not contain comment yet, but they'll be there soon
2. Added some funcionality to music class: now by default it tries to
catch as much information as possible from the file:
a. title: if it can't get it the class uses file name without last four
characters (.mp3) as the title
b. artist: if it can't get it the class leaves that field blank
c. category: checks if genre name contains items in catList, if it does
method sets category value to catList item, if it does not find a macth
method uses first element on the catList - "unknown"
P. S. if catList item contains two or more words (hip hop) method does
checking only on first word
d. time: tries to get time if it is eaquals to 0 it sets as "unknown"
P. S. I did not get any exceptions here so there the statement which get
time (check the code and you'll see) is not surrounded with try-catch.
Time seems to be good, although I found one  song on my playlist that
returned time value of 0, nevertheless it seems to be playing well

I haven't tested this player on other formats than mp3 and file filter
now reads only mp3 files